backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0
openai>=1.12.0
alembic>=1.13.0
celery[redis]>=5.3.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
scikit-learn>=1.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
sse-starlette>=2.0.0
